refactor : Logger 리뷰 반영
