+++ src/main/java/com/woowacourse/pelotonbackend/common/GlobalExceptionHandler.java
@@ -20,7 +20,7 @@ public class GlobalExceptionHandler {
 
     @ExceptionHandler({MethodArgumentNotValidException.class, MethodArgumentTypeMismatchException.class})
     protected ResponseEntity<ErrorResponse> validException(final MethodArgumentNotValidException exception) {
-        log.error("Validate Exception ! : {} ", exception.toString(), exception);
+        log.info("Validate Exception ! : {} ", exception.toString(), exception);
 
         final ErrorCode errorCode = ErrorCode.INVALID_VALIDATE;
         final ErrorResponse errorResponse = ErrorResponse.of(errorCode.getStatus(), errorCode.getCode(),
@@ -31,7 +31,7 @@ public class GlobalExceptionHandler {
 
     @ExceptionHandler(BindException.class)
     protected ResponseEntity<ErrorResponse> bindException(final BindException exception) {
-        log.error("HandleBind Exception ! : {} ", exception.toString(), exception);
+        log.info("HandleBind Exception ! : {} ", exception.toString(), exception);
 
         final ErrorCode errorCode = ErrorCode.INVALID_VALIDATE;
         final ErrorResponse errorResponse = ErrorResponse.of(
@@ -45,7 +45,7 @@ public class GlobalExceptionHandler {
 
     @ExceptionHandler({ConstraintViolationException.class, HttpMessageNotReadableException.class})
     protected ResponseEntity<ErrorResponse> invalidInputException(final RuntimeException exception) {
-        log.error(exception.getClass().getSimpleName() + " Exception !", exception);
+        log.info(exception.getClass().getSimpleName() + " Exception !", exception);
 
         final ErrorCode errorCode = ErrorCode.INVALID_VALIDATE;
         final ErrorResponse errorResponse = ErrorResponse.of(errorCode.getStatus(), errorCode.getCode(),
@@ -56,7 +56,7 @@ public class GlobalExceptionHandler {
 
     @ExceptionHandler(BusinessException.class)
     protected ResponseEntity<ErrorResponse> businessException(final BusinessException exception) {
-        log.error("Business Exception ! : {}", exception.toString(), exception);
+        log.info("Business Exception ! : {}", exception.toString(), exception);
 
         final ErrorCode errorCode = exception.getErrorCode();
         final ErrorResponse errorResponse = ErrorResponse.of(
