+++ src/main/java/com/woowacourse/pelotonbackend/config/FilterConfig.java
@@ -1,8 +1,10 @@
-package com.woowacourse.pelotonbackend.support.log;
+package com.woowacourse.pelotonbackend.config;
 
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+
+import com.woowacourse.pelotonbackend.support.log.HttpLoggerFilter;
 
 @Configuration
 public class FilterConfig {
