+++ src/main/resources/db/changelog-master.xml
@@ -4,5 +4,4 @@
         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
         xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.8.xsd">
     <include file="./changes/changelog-schema-1.0.0.xml" relativeToChangelogFile="true"/>
-    <include file="./changes/changelog-schema-1.0.1.xml" relativeToChangelogFile="true"/>
 </databaseChangeLog>
