--- src/main/resources/db/changes/changelog-schema-1.0.1.xml
@@ -1,12 +1,0 @@
-<?xml version="1.1" encoding="UTF-8" standalone="no"?>
-<databaseChangeLog xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                   xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
-                   xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.5.xsd">
-
-    <changeSet author="kyle" id="20200805">
-         <sql  dbms="h2, mysql" stripComments="true">
-         ALTER TABLE MEMBER ALTER COLUMN PROFILE_BASE_IMAGE_URL CLOB NULL;
-        <comment>change profile not null to null</comment>
-    </sql>
-    </changeSet>
-</databaseChangeLog>
