+++ src/main/resources/log4j2-prod.xml
@@ -6,6 +6,7 @@
             ERROR=red,INFO=green, DEBUG=blue} [%c] %style{[%t]}{yellow}- %m%n -
         </Property>
     </Properties>
+
     <Appenders>
         <RollingFile name="ErrorAppender" fileName="logs/error/${logNm}.log"
                      filePattern="logs/error/${logNm}_%d{yyyy-MM-dd}_%i.log">
@@ -38,18 +39,22 @@
             <DefaultRolloverStrategy max="30" fileIndex="max"/>
         </RollingFile>
     </Appenders>
+
     <Loggers>
-        <Root level="ERROR" includeLocation="true" additivity="false">
-            <AppenderRef ref="ErrorAppender" />
+        <Root level="ERROR" includeLocation="false" additivity="false">
+            <AppenderRef ref="ErrorAppender"/>
         </Root>
-        <Logger name="com.woowacourse.pelotonbackend.support.log.HttpLoggerFilter" level="INFO" additivity="false">
-            <AppenderRef ref="HttpAppender" />
+
+        <Logger name="com.woowacourse.pelotonbackend.support.log.HttpLoggerFilter" level="INFO" additivity="false" includeLocation="false">
+            <AppenderRef ref="HttpAppender"/>
         </Logger>
-        <Logger name="org.springframework.jdbc" level="DEBUG" additivity="true">
-            <AppenderRef ref="SqlAppender" />
+
+        <Logger name="org.springframework.jdbc" level="DEBUG" additivity="true" includeLocation="false">
+            <AppenderRef ref="SqlAppender"/>
         </Logger>
-        <Logger name="com.zaxxer.hikari.HikariConfig" level="DEBUG" additivity="true">
-            <AppenderRef ref="SqlAppender" />
+
+        <Logger name="com.zaxxer.hikari.HikariConfig" level="DEBUG" additivity="true" includeLocation="false">
+            <AppenderRef ref="SqlAppender"/>
         </Logger>
     </Loggers>
 </Configuration>
