+++ src/main/resources/log4j2-secret.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<Configuration status="warn">
+<Configuration status="WARN">
     <Properties>
         <Property name="logNm">peloton-log</Property>
         <Property name="layoutPattern">%style{%d{yyyy/MM/dd HH:mm:ss,SSS}}{blue} %highlight{[%p]}{FATAL=bg_red,
